- Se corrigen los tipos de accountNumber e identifier para pasarlos de integer a long.
- Se cambia path de eliminación de cuenta de terceros para unificar path entre endpoints de eliminación de cuentas asociadas.

Cherry picked from !11593

+++ bff/src/main/java/bg/com/bo/bff/providers/implementations/AchAccountMiddlewareProvider.java
@@ -108,7 +108,7 @@ public class AchAccountMiddlewareProvider implements IAchAccountProvider {
     }
 
     @Override
-    public GenericResponse deleteAchAccount(String personId, int identifier, String deviceId, String deviceIp) throws IOException {
+    public GenericResponse deleteAchAccount(String personId, long identifier, String deviceId, String deviceIp) throws IOException {
         ClientToken clientToken = tokenMiddlewareProvider.generateAccountAccessToken(ProjectNameMW.ACH_ACCOUNTS.getName(), middlewareConfig.getClientAchAccount(), ProjectNameMW.ACH_ACCOUNTS.getHeaderKey());
         DeleteAchAccountMWRequest requestData = mapper.convert(personId, identifier);
